More tweaks to gutters & spacing. Fixed issue with permalinks. Set summary list to vertical bullets. Removed redundant html from template files. Added widget support. Segregated Import & Settings tab code to their own separate files.
