+++ includes/admin_import.php
@@ -454,14 +454,29 @@ if (!function_exists('tsml_import_page')) {
 		<style>
 			.column {
 			  vertical-align: top;
-			  padding: 10px 0px;
+			  padding: 20px 0 0 0;
 			  display: inline-block;
 			  width: 100%;
 			}
 
+			#import_settings_wrap .settings-card {
+				padding: 20px;
+				margin: 10px;
+				line-height: 100%;
+				height: auto;
+			}
+
+			#import_settings_wrap .import-card {
+				margin: 10px;
+				margin-top: 20px;
+				padding: 0 20px 0 20px;
+				line-height: 100%;
+				height: auto;
+			}
+
 			@media only screen and (min-width: 900px) {
-			.column.one-third { width: 33%; }
-			.column.three-third { width:100%; }
+			.column.one-third { width: 33.2%; }
+			.column.three-third { width:99.6%; }
 			}
 
 			.column div {
@@ -469,7 +484,12 @@ if (!function_exists('tsml_import_page')) {
 			}
 
 			.column p {
-			  margin: 0px;
+			  margin: 10px;
+			}
+
+			.no_space_after p {
+			  margin: 0;
+			  padding: 0;
 			}
 
 			#about-us .left {
@@ -477,7 +497,8 @@ if (!function_exists('tsml_import_page')) {
 			  width: 60%;
 			}
 			#about-us .middle {
-			  float: left;
+			  padding: 20px;
+			  width: 25px;
 			}
 			#about-us .right {
 			  float: right;
@@ -490,40 +511,11 @@ if (!function_exists('tsml_import_page')) {
 			  margin: 0;
 			}
 
-			.shadow {
-			  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2); 
-			}
-
-			#import_settings_wrap .settings-card {
-				padding:10px;
-				line-height: 50%;
-				height: auto;
-			}
-
 			#import_settings_wrap .li .p {
 				margin: 0;
-				padding: 0;
+				padding: 10px;
 			}
 
-			#import_settings_wrap .import-card {
-				padding:10px;
-				line-height: 50%;
-				height: 300px;
-			}
-
-			#import-1 .li .p {
-				margin: 0;
-				padding: 0;
-			}
-
-			#import-data-source {
-				margin-right: 20px;
-				margin-left: 10px;
-				padding-right: 20px;
-			}
-
-			#ul_horiz_summary li { list-style: disc; padding: 1px 20px; display: inline; margin:0 auto;  border: solid; border-width: 1px; border-color: lightgray; }
-		
 
 		</style>
 
@@ -614,329 +606,14 @@ if (!function_exists('tsml_import_page')) {
 
 			<div class="tab-content">
 			<?php switch($tab) :
-			case 'settings':   ?>
+			case 'settings':   
 
-				<!-- Settings tab HTML goes here -->
-				<div id="settings-row1" class="" >
+				$file = TSML_PATH . "includes/admin_settings_tab.php";
 
-					<div id="column1" class="column one-third" >
-						<!-- Put General Settings section here -->
-						<div class="postbox " >
-							<div class="inside">
-								<h1><?php _e('General', '12-step-meeting-list') ?></h1>
-								<form method="post" class="settings-card" action="<?php echo $_SERVER['REQUEST_URI'] ?>">
-										<h3><strong><?php _e('Program', '12-step-meeting-list') ?></strong></h3>
-										<p><?php _e('Select the Recovery Program your site targets here.', '12-step-meeting-list') ?></p>
-									<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false) ?>
-									<select name="tsml_program" onchange="this.form.submit()">
-										<?php foreach ($tsml_programs as $key => $value) { ?>
-											<option value="<?php echo $key ?>" <?php selected($tsml_program, $key) ?>><?php echo $value['name'] ?></option>
-										<?php } ?>
-									</select>
-								</form>
-								<form method="post" class="settings-card" action="<?php echo $_SERVER['REQUEST_URI'] ?>">
-										<h3><strong><?php _e('Distance Units', '12-step-meeting-list') ?></strong></h3>
-										<p><?php _e('This determines which units are used on the meeting list page.', '12-step-meeting-list') ?></p>
-									<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false) ?>
-									<select name="tsml_distance_units" onchange="this.form.submit()">
-										<?php
-										foreach ([
-											'km' => __('Kilometers', '12-step-meeting-list'),
-											'mi' => __('Miles', '12-step-meeting-list'),
-										] as $key => $value) { ?>
-											<option value="<?php echo $key ?>" <?php selected($tsml_distance_units, $key) ?>>
-												<?php echo $value ?>
-											</option>
-										<?php } ?>
-									</select>
-								</form>
-								<form method="post" class="settings-card" action="<?php echo $_SERVER['REQUEST_URI'] ?>">
-										<h3><strong><?php _e('Contact Visibility', '12-step-meeting-list') ?></strong></h3>
-										<p><?php _e('This determines whether contacts are displayed publicly on meeting detail pages.', '12-step-meeting-list') ?></p>
-									<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false) ?>
-									<select name="tsml_contact_display" onchange="this.form.submit()">
-										<?php
-										foreach ([
-											'public' => __('Public', '12-step-meeting-list'),
-											'private' => __('Private', '12-step-meeting-list'),
-										] as $key => $value) { ?>
-											<option value="<?php echo $key ?>" <?php selected($tsml_contact_display, $key) ?>>
-												<?php echo $value ?>
-											</option>
-										<?php } ?>
-									</select>
-								</form>
-							</div>
-						</div>
-
-						<div class="postbox " >
-						<!-- Put Feed Management here -->
-							<div class="inside">
-								<h1><?php _e('Feed Management', '12-step-meeting-list') ?></h1>
-								<form method="post" class="settings-card" action="<?php echo $_SERVER['REQUEST_URI'] ?>">
-									<h3><strong><?php _e('Feed Sharing', '12-step-meeting-list') ?></strong></h3>
-									<p><?php printf(__('Open means your feeds are available publicly. Restricted means people need a key or to be logged in to get the feed.', '12-step-meeting-list')) ?></p>
-									<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false) ?>
-									<select name="tsml_sharing" onchange="this.form.submit()">
-										<?php
-										foreach ([
-											'open' => __('Open', '12-step-meeting-list'),
-											'restricted' => __('Restricted', '12-step-meeting-list'),
-										] as $key => $value) { ?>
-											<option value="<?php echo $key ?>" <?php selected($tsml_sharing, $key) ?>>
-												<?php echo $value ?>
-											</option>
-										<?php } ?>
-									</select>
-								</form>
-									<?php if ($tsml_sharing == 'restricted') { ?>
-										<h3><strong><?php _e('Authorized Apps', '12-step-meeting-list') ?></strong></h3>
-										<p><?php _e('You may allow access to your meeting data for specific purposes, such as the <a target="_blank" href="https://meetingguide.org/">Meeting Guide App</a>.', '12-step-meeting-list') ?></p>
-
-										<?php if (count($tsml_sharing_keys)) { ?>
-											<table class="tsml_sharing_list">
-												<?php foreach ($tsml_sharing_keys as $key => $name) {
-													$address = admin_url('admin-ajax.php?') . http_build_query([
-														'action' => 'meetings',
-														'key' => $key,
-													]);
-												?>
-													<tr>
-														<td><a href="<?php echo $address ?>" target="_blank"><?php echo $name ?></a></td>
-														<td>
-															<form method="post" action="<?php echo $_SERVER['REQUEST_URI'] ?>">
-																<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false) ?>
-																<input type="hidden" name="tsml_remove_sharing_key" value="<?php echo $key ?>">
-																<span class="dashicons dashicons-no-alt"></span>
-															</form>
-														</td>
-													</tr>
-												<?php } ?>
-											</table>
-										<?php } ?>
-										<form class="columns" method="post" action="<?php echo $_SERVER['REQUEST_URI'] ?>">
-									<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false) ?>
-									<div class="input">
-										<input type="text" name="tsml_add_sharing_key" placeholder="<?php _e('Meeting Guide', '12-step-meeting-list') ?>">
-									</div>
-									<div class="btn">
-										<input type="submit" class="button" value="<?php _e('Add', '12-step-meeting-list') ?>">
-									</div>
-								</form>
-							<?php } else { ?>
-								<h3><strong><?php _e('Public Feed', '12-step-meeting-list') ?></strong></h3>
-								<p><?php _e('The following feed contains your publicly available meeting information.', '12-step-meeting-list') ?></p>
-								<?php printf(__('<a class="public_feed" href="%s" target="_blank">Public Data Source</a>', '12-step-meeting-list'), admin_url('admin-ajax.php?action=meetings')) ?>
-							<?php } ?>
-
-							</div>
-						</div>
-					</div>
-
-					<div id="column2" class="column one-third "  >
-						<div class="postbox " >
-						<!-- Put Switch UI here -->
-							<div class="inside">
-								<h1><?php _e('Switch UI', '12-step-meeting-list') ?></h1>
-								<div class="settings-card" >
-									<h3><strong><?php _e('User Interface Display', '12-step-meeting-list') ?></strong></h3>
-									<p><?php _e('Please select the user interface design that is right for your site. Choose between our latest design that we call <b>TSML UI</b> or stay with the old standard <b>Legacy UI</b>.', '12-step-meeting-list') ?></p>							
-									<form method="post" action="<?php echo $_SERVER['REQUEST_URI'] ?>">
-										<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false) ?>
-										<select name="tsml_user_interface" onchange="this.form.submit()">
-											<option value="legacy" <?php selected($tsml_user_interface, 'legacy_ui') ?> >
-												<?php _e('Legacy UI', '12-step-meeting-list') ?>
-											</option>
-											<option value="tsml_ui" <?php selected($tsml_user_interface, 'tsml_ui') ?> >
-												<?php _e('TSML UI', '12-step-meeting-list') ?>
-											</option>
-										</select>
-									</form>
-								</div>
-							</div>
-						</div>
-
-						<!-- Put Map Settings here -->
-						<div class="postbox " >
-							<div class="inside">
-								<h1><?php _e('Mapping & Geocoding', '12-step-meeting-list') ?></h1>
-								<p style="padding-left:20px;"><?php _e('Display of maps requires an authorization key from </strong><a href="https://www.mapbox.com/" target="_blank">Mapbox</a></strong> or </strong><a href="https://console.cloud.google.com/home/" target="_blank">Google</a></strong>.', '12-step-meeting-list') ?></p>
-
-								<div class="settings-card">
-									<h3><strong><?php _e('Mapbox Access Token', '12-step-meeting-list') ?></strong></h3>
-									<p><?php _e('Enter a key from Mapbox to enable their Web Map Services.', '12-step-meeting-list') ?></p>
-									<!-- Put Mapbox Signup Instructions here -->
-									<?php if (empty($tsml_mapbox_key)) : ?>
-										<div class="location_note">
-											<p>To get signed up for the <strong>Mapbox Map Service </strong><a href="https://www.mapbox.com/signup/" target="_blank">Go Here</a>. You will only need
-												a valid email address. <i>No credit card is required</i>. Once signed up you can copy and paste your access token below.</p>
-											<p>*Please note: TSML UI only supports Mapbox, not Google.</p>
-										</div><br>
-										<?php endif; ?>
-										<form class="columns" method="post" action="<?php echo $_SERVER['REQUEST_URI'] ?>">
-											<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false) ?>
-											<div class="input">
-												<input type="text" name="tsml_add_mapbox_key" value="<?php echo $tsml_mapbox_key ?>" placeholder="Enter Mapbox access token here">
-											</div>
-											<div class="btn">
-												<?php if (empty($tsml_mapbox_key)) { ?>
-													<input type="submit" class="button" value="<?php _e('Add', '12-step-meeting-list') ?>">
-												<?php } else { ?>
-													<input type="submit" class="button" value="<?php _e('Update', '12-step-meeting-list') ?>">
-												<?php } ?>
-											</div>
-										</form>
-									</div>
-								<div class="settings-card">
-									<h3><strong><?php _e('Google Maps API Key', '12-step-meeting-list') ?></strong></h3>
-									<p><?php _e('Enter a key from Google authorizing use of their Map Services (<a href="https://developers.google.com/maps/documentation/javascript/get-api-key" target="_blank">get instructions here</a>).', '12-step-meeting-list') ?></p>
-									<form class="columns" method="post" action="<?php echo $_SERVER['REQUEST_URI'] ?>">
-										<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false) ?>
-										<div class="input">
-											<input type="text" name="tsml_add_google_maps_key" value="<?php echo $tsml_google_maps_key ?>" placeholder="Enter Google API key here">
-										</div>
-										<div class="btn">
-											<?php if (empty($tsml_google_maps_key)) { ?>
-												<input type="submit" class="button" value="<?php _e('Add', '12-step-meeting-list') ?>">
-											<?php } else { ?>
-												<input type="submit" class="button" value="<?php _e('Update', '12-step-meeting-list') ?>">
-											<?php } ?>
-										</div>
-									</form>
-								</div>
-							
-								<div class="settings-card">
-									<h3><strong><?php _e('Address Geocoding', '12-step-meeting-list') ?></strong></h3>
-									<p><?php _e('Code4Recovery is working on a new method for geocoding addresses. You can decide whether to use the old way, or whether to help us test the new way.', '12-step-meeting-list') ?></p>
-								<form method="post" action="<?php echo $_SERVER['REQUEST_URI'] ?>">
-									<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false) ?>
-									<select name="tsml_geocoding_method" onchange="this.form.submit()">
-										<option value="legacy" <?php selected($tsml_geocoding_method, 'legacy') ?>>
-											<?php _e('Legacy Method', '12-step-meeting-list') ?>
-										</option>
-										<?php if (!empty($tsml_google_maps_key)) { ?>
-											<option value="google_key" <?php selected($tsml_geocoding_method, 'google_key') ?>>
-												<?php _e('Use my Google API Key', '12-step-meeting-list') ?>
-											</option>
-										<?php } ?>
-										<option value="api_gateway" <?php selected($tsml_geocoding_method, 'api_gateway') ?>>
-											<?php _e('BETA - API Gateway - BETA', '12-step-meeting-list') ?>
-										</option>
-									</select>
-								</form>
-								<?php if (!empty($tsml_google_maps_key)) { ?>
-									<p>If you select "Use my Google API Key", then you <strong>must</strong> go into the Google Console and enable the geocode API for your key</p>
-								<?php } ?>
-								</div>
-							</div>
-						</div>
-					</div>
-
-					<div id="column3" class="column one-third" >
-						<!-- Put About Us here -->
-						<div class="postbox" >
-							<div class="inside">
-								<div class="inner-column" style="padding: 0; margin: 0;">
-									<div class="right" style="float:right;  padding: 0; margin:20px auto;" >
-										<a href="https://code4recovery.org"><img src="/wp-content/plugins/12-step-meeting-list/assets/img/code4recovery.svg" alt="Code For Recovery" ></a>										
-									</div>
-									<div class="left">
-										<h1  style="padding-left:0; "><?php _e('About Us', '12-step-meeting-list') ?></h1>
-										<p style="padding: 20px; 10px; 0 10px;"><?php _e('This <b>12 Step Meeting List</b> plugin (TSML) is one of the free services offered by the nonprofit organization <b>Code For Recovery</b> whose volunteer members build and maintain technology services for recovery fellowships such as AA and Al-Anon.', '12-step-meeting-list') ?></p>										
-									</div>
-								</div>
-							</div>
-						</div>
-
-						<div class="postbox " >
-						<!-- Put Need Help here -->
-							<div class="inside">
-								<h1><?php _e('Need Help?', '12-step-meeting-list') ?></h1>
-								<div class="settings-card">
-									
-									<p><?php _e("To get information about this product or our organization, simply use one of the linked buttons below. Both the plugin Wiki page and/or the <b>Code For Recovery</b> website are great sources of information. You can also ask questions directly throght our GitHub discussion forum which is monitored daily by members of our maintenance team.", '12-step-meeting-list') ?></p>
-									<div style="margin:10px;">
-										<a href="https://github.com/code4recovery/12-step-meeting-list/wiki/" target="_blank" class="button" style=" margin-right: 35px;">
-											<?php _e('Go to our Wiki', '12-step-meeting-list') ?>
-										</a> 
-										<a href="https:///code4recovery.org/" target="_blank" class="button" style="margin-right: 35px;">
-											<?php _e('Code For Recovery website', '12-step-meeting-list') ?>
-										</a>
-										<a href="https://github.com/code4recovery/12-step-meeting-list/discussions" target="_blank" class="button" style="">
-											<?php _e('Ask a Question', '12-step-meeting-list') ?>
-										</a> 
-									</div>
-								</div>
-							</div>
-						</div>
-
-						<!-- Put Email Settings here -->
-						<div class="postbox" >
-							<div class="inside" >
-								<h1><?php _e('Email Addresses', '12-step-meeting-list') ?></h1>
-								<div class="settings-card" >
-										<h3><strong><?php _e('User Feedback Emails', '12-step-meeting-list') ?></strong></h3>
-										<p><?php _e('Enable a meeting info feedback form by adding email addresses here.', '12-step-meeting-list') ?></p>
-
-									<?php if (!empty($tsml_feedback_addresses)) { ?>
-										<table class="tsml_address_list">
-											<?php foreach ($tsml_feedback_addresses as $address) { ?>
-												<tr>
-													<td><?php echo $address ?></td>
-													<td>
-														<form method="post" action="<?php echo $_SERVER['REQUEST_URI'] ?>">
-															<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false) ?>
-															<input type="hidden" name="tsml_remove_feedback_address" value="<?php echo $address ?>">
-															<span class="dashicons dashicons-no-alt"></span>
-														</form>
-													</td>
-												</tr>
-											<?php } ?>
-										</table>
-									<?php } ?>
-									<form class="columns" method="post" action="<?php echo $_SERVER['REQUEST_URI'] ?>">
-										<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false) ?>
-										<div class="input">
-											<input type="email" name="tsml_add_feedback_address" placeholder="email@example.org">
-										</div>
-										<div class="btn">
-											<input type="submit" class="button" value="<?php _e('Add', '12-step-meeting-list') ?>">
-										</div>
-									</form>
-								</div>
-								<div class="settings-card" >
-									<h3><strong><?php _e('Change Notification Emails', '12-step-meeting-list') ?></strong></h3>
-									<p><?php _e('Receive notifications of meeting changes at the email addresses below.', '12-step-meeting-list') ?></p>
-									<?php if (!empty($tsml_notification_addresses)) { ?>
-										<table class="tsml_address_list">
-											<?php foreach ($tsml_notification_addresses as $address) { ?>
-												<tr>
-													<td><?php echo $address ?></td>
-													<td>
-														<form method="post" action="<?php echo $_SERVER['REQUEST_URI'] ?>">
-															<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false) ?>
-															<input type="hidden" name="tsml_remove_notification_address" value="<?php echo $address ?>">
-															<span class="dashicons dashicons-no-alt"></span>
-														</form>
-													</td>
-												</tr>
-											<?php } ?>
-										</table>
-									<?php } ?>
-									<form class="columns" method="post" action="<?php echo $_SERVER['REQUEST_URI'] ?>">
-										<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false) ?>
-										<div class="input">
-											<input type="email" name="tsml_add_notification_address" placeholder="email@example.org">
-										</div>
-										<div class="btn">
-											<input type="submit" class="button" value="<?php _e('Add', '12-step-meeting-list') ?>">
-										</div>
-									</form>
-								</div>
-							</div>
-						</div>
-					</div>
-				</div>
+				if (file_exists($file)) {
+					require $file;
+				}
+				?>
 
 			  <?php break;
 			case 'example':   ?>
@@ -985,241 +662,13 @@ if (!function_exists('tsml_import_page')) {
 
 			  <?php break;
 
-			default:  ?>
-				<!-- Import HTML goes here -->
-				<div id="import-1" class="">
+			default:  
+				$file = TSML_PATH . "includes/admin_import_tab.php";
 
-					<div id="col1" class="column one-third" >
-						<!-- Put Import CSV section here -->
-						<div id="import-csv" class="postbox  import-card" >
-							<div class="inside">
-
-								<h3><?php _e('Import CSV', '12-step-meeting-list') ?></h3>
-								<form method="post" action="<?php echo $_SERVER['REQUEST_URI'] ?>" enctype="multipart/form-data">
-									<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false) ?>
-									<input type="file" name="tsml_import">
-									<p>
-										<?php _e('When importing...', '12-step-meeting-list') ?><br>
-										<?php
-										$delete_options = [
-											'nothing'	=> __('don\'t delete anything', '12-step-meeting-list'),
-											'regions'	=> __('delete only the meetings, locations, and groups for the regions present in this CSV', '12-step-meeting-list'),
-											'all' 		=> __('delete all meetings, locations, groups, districts, and regions', '12-step-meeting-list'),
-										];
-										if (!empty($tsml_data_sources)) {
-											$delete_options['no_data_source'] = __('delete all meetings, locations, and groups not from a data source', '12-step-meeting-list');
-										}
-										$delete_selected = (empty($_POST['delete']) || !array_key_exists($_POST['delete'], $delete_options)) ? 'nothing' : $_POST['delete'];
-										foreach ($delete_options as $key => $value) { ?>
-											<label><input type="radio" name="delete" value="<?php echo $key ?>" <?php checked($key, $delete_selected) ?>> <?php echo $value ?></label><br>
-										<?php } ?>
-									</p>
-									<p><input type="submit" class="button button-primary" value="<?php _e('Begin', '12-step-meeting-list') ?>"></p>
-								</form>
-							</div>
-						</div>
-					</div>
-
-					<div id="col2" class="column one-third " >
-						<!-- Put Wheres My Info? section here -->
-						<div id="wheres_my_info" class="postbox import-card">
-							<div class="inside">
-								<?php
-								$meetings = tsml_count_meetings();
-								$locations = tsml_count_locations();
-								$regions = tsml_count_regions();
-								$groups = tsml_count_groups();
-
-								$pdf_link = 'https://pdf.code4recovery.org/?' . http_build_query([
-									'json' => admin_url('admin-ajax.php') . '?' . http_build_query([
-										'action' => 'meetings',
-										'nonce' => $tsml_sharing === 'restricted' ? wp_create_nonce($tsml_nonce) : null
-									])
-								]);
-
-								?>
-								<h3><?php _e('Where\'s My Info?', '12-step-meeting-list') ?></h3>
-								<?php if ($tsml_slug) { ?>
-									<p><?php printf(__('Your public meetings page is <a href="%s">right here</a>. Link that page from your site\'s nav menu to make it visible to the public.', '12-step-meeting-list'), get_post_type_archive_link('tsml_meeting')) ?></p>
-								<?php
-								}
-								if ($meetings) { ?>
-									<p><?php printf(__('<strong>Going away soon:</strong> a very basic PDF schedule is available in three sizes: <a href="%s">4&times;7</a>, <a href="%s">half page</a> and <a href="%s">full page</a>.', '12-step-meeting-list'), admin_url('admin-ajax.php') . '?action=tsml_pdf&width=4&height=7', admin_url('admin-ajax.php') . '?action=tsml_pdf', admin_url('admin-ajax.php') . '?action=tsml_pdf&width=8.5') ?></p>
-									<p><?php _e('<strong>New!</strong> We are developing a service to generate PDF directories of in-person meetings.', '12-step-meeting-list') ?></p>
-									<p><a href="<?php echo $pdf_link ?>" target="_blank" class="button"><?php _e('Generate PDF') ?></a>
-								<?php } ?>
-
-								<div id="tsml_counts" <?php if (($meetings + $locations + $groups + $regions) == 0) { ?> class="hidden" <?php } ?> >
-									<p><?php _e('You have:', '12-step-meeting-list') ?></p>
-									<div class="table">
-										<ul id="ul_horiz_summary" class="" style="width:100%;">
-											<li class="meetings<?php if (!$meetings) { ?> hidden<?php } ?>">
-												<?php printf(_n('%s meeting', '%s meetings', $meetings, '12-step-meeting-list'), number_format_i18n($meetings)) ?>
-											</li>
-											<li class="locations<?php if (!$locations) { ?> hidden<?php } ?>">
-												<?php printf(_n('%s location', '%s locations', $locations, '12-step-meeting-list'), number_format_i18n($locations)) ?>
-											</li>
-											<li class="groups<?php if (!$groups) { ?> hidden<?php } ?>">
-												<?php printf(_n('%s group', '%s groups', $groups, '12-step-meeting-list'), number_format_i18n($groups)) ?>
-											</li>
-											<li class="regions<?php if (!$regions) { ?> hidden<?php } ?>">
-												<?php printf(_n('%s region', '%s regions', $regions, '12-step-meeting-list'), number_format_i18n($regions)) ?>
-											</li>
-										</ul>
-									</div>
-								</div>	
-							</div>
-						</div>
-					</div>
-			
-					<div id="col3" class="column one-third"  >
-						<!-- Put Export Meeting List section here -->
-						<div id="export_meeting_list" class="postbox import-card"  >
-							<div class="inside" >
-								<h3><?php _e('Export Meeting List', '12-step-meeting-list') ?></h3>
-								<?php
-								if ($meetings) { ?>
-									<p><?php printf(__('You can download your meetings in <a href="%s">CSV format</a>.', '12-step-meeting-list'), admin_url('admin-ajax.php') . '?action=csv') ?></p>
-								<?php } ?><br>
-								<p><?php printf(__('Want to send a mass email to your contacts? <br><a href="%s" target="_blank">Click here</a> to see their email addresses.', '12-step-meeting-list'), admin_url('admin-ajax.php') . '?action=contacts') ?></p>
-							</div>
-						</div>
-					</div>
-				</div>
-
-				<div id="import-2" class="">
-					<!-- Put Import Data Sources section here -->
-					<div id="import-data-source" class="postbox " >
-						<div class="inside column three-third" style="">
-							<h3><?php _e('Import Data Sources', '12-step-meeting-list')?></h3>
-							<p style="padding: 0 20px 0 20px;">
-							<?php printf(__('Data sources are JSON feeds that contain a website\'s public meeting data. They can be used to aggregate meetings from different sites into a single master list. 
-								Data sources listed below will pull meeting information into this website. A configurable schedule allows for each enabled data source to be scanned at least once per day looking 
-								for updates to the listing. Change Notification email addresses are sent an email when action is required to re-sync a data source with its meeting list information. 
-								Please note: records that you intend to maintain on your website should always be imported using the Import CSV feature above. <b>Data Source records will be overwritten when the 
-								parent data source is refreshed. </b>More information is available at the <a href="%s" target="_blank">Meeting Guide API Specification</a>.', '12-step-meeting-list'), 'https://github.com/code4recovery/spec')?></p> 
-							<?php if (!empty($tsml_data_sources)) { ?>
-							<table>
-								<thead>
-									<tr>
-										<th class="small"></th>
-										<th class=""><?php _e('Feed', '12-step-meeting-list') ?> </th>
-										<th class="align-left"><?php _e('Parent Region', '12-step-meeting-list') ?></th>
-										<th class="align-left"><?php _e('Change Detection', '12-step-meeting-list') ?></th>
-										<th class="align-center"><?php _e('Meetings', '12-step-meeting-list') ?></th>
-										<th class="align-right"><?php _e('Last Refresh', '12-step-meeting-list') ?></th>
-										<th class="small"></th>
-									</tr>
-								</thead>
-								<tbody>
-									<?php foreach ($tsml_data_sources as $feed => $properties) { ?>
-									<tr data-source="<?php echo $feed?>">
-										<td class="small ">
-											<form method="post" action="<?php echo $_SERVER['REQUEST_URI']?>">
-												<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false)?>
-												<input type="hidden" name="tsml_add_data_source" value="<?php echo $feed?>">
-												<input type="hidden" name="tsml_add_data_source_name" value="<?php echo @$properties['name']?>">
-												<input type="hidden" name="tsml_add_data_source_parent_region_id" value="<?php echo @$properties['parent_region_id']?>">
-												<input type="hidden" name="tsml_add_data_source_change_detect" value="<?php echo @$properties['change_detect'] ?>">
-												<input type="submit" value="Refresh" class="button button-small">
-											</form>
-										</td>
-										<td>
-											<a href="<?php echo $feed?>" target="_blank">
-												<?php echo !empty($properties['name']) ? $properties['name'] : __('Unnamed Feed', '12-step-meeting-list')?>
-											</a>
-										</td>
-										<td>
-											<?php
-												$parent_region = null;
-												if (empty($properties['parent_region_id']) || $properties['parent_region_id'] == -1) {
-													$parent_region = __('Top-level region', '12-step-meeting-list');
-												} elseif (empty($regions[$properties['parent_region_id']])) {
-													$term = get_term_by('term_id', $properties['parent_region_id'], 'tsml_region');
-													$parent_region = $term->name;
-													if ($parent_region == null) {
-														$parent_region = 'Missing Parent Region: ' . $properties['parent_region_id'];
-													}
-												} else {
-													$parent_region = $regions[$properties['parent_region_id']];
-												}
-												echo $parent_region;
-											?>
-										</td>
-										<td>
-											<?php
-												$change_detect = null;
-												if (empty($properties['change_detect']) || $properties['change_detect'] == -1) {
-													$change_detect = __('Disabled', '12-step-meeting-list');
-												} else {
-													$change_detect = ucfirst($properties['change_detect']);
-												}
-
-												echo $change_detect;
-											?>
-										</td>
-										<td class="align-center count_meetings"><?php echo number_format($properties['count_meetings']) ?></td>
-
-										<td class="align-right">
-											<?php echo Date(get_option('date_format') . ' ' . get_option('time_format'), $properties['last_import'])?>
-										</td>
-
-										<td class="small">
-											<form method="post" action="<?php echo $_SERVER['REQUEST_URI']?>">
-												<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false)?>
-												<input type="hidden" name="tsml_remove_data_source" value="<?php echo $feed?>">
-												<span class="dashicons dashicons-no-alt"></span>
-											</form>
-										</td>
-									</tr>
-									<?php }?>
-								</tbody>
-							</table>
-							<?php } ?>
-							<form class="columns" method="post" action="<?php echo $_SERVER['REQUEST_URI']?>">
-								<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false)?>
-
-								<div>
-									<input type="text" name="tsml_add_data_source_name" placeholder="<?php _e('District 02', '12-step-meeting-list')?>">
-								</div>
-
-								<div class="input-half">
-									<input type="text" name="tsml_add_data_source" placeholder="https://">
-								</div>
-
-								<div class="input-data-source input-region">
-									<?php wp_dropdown_categories(array(
-										'name' => 'tsml_add_data_source_parent_region_id',
-										'taxonomy' => 'tsml_region',
-										'hierarchical' => true,
-										'hide_empty' => false,
-										'orderby' => 'name',
-										'selected' => null,
-										'title' => __('Append regions created by this data source to… (top-level, if none selected)', '12-step-meeting-list'),
-										'show_option_none' => __('Parent Region…', '12-step-meeting-list'),
-									))?>
-								</div>
-
-								<div class="input-data-source input-auto-refresh" >
-									<?php wp_nonce_field($tsml_nonce, 'tsml_nonce', false)?>
-									<select name="tsml_add_data_source_change_detect" id="tsml_change_detect" >
-									<?php 
-									foreach (array(
-											'disabled' => __('Change Detection Disabled', '12-step-meeting-list'),
-											'enabled' => __('Change Detection Enabled', '12-step-meeting-list'),	
-										) as $key => $value) {?>
-										<option value="<?php echo $key?>"<?php selected($tsml_change_detect, $key) ?> ><?php echo $value?></option>
-									<?php } ?>
-									</select>
-								</div>
-								<div>
-									<input type="submit" class="btn button" value="<?php _e('Add Data Source', '12-step-meeting-list') ?>">
-								</div>
-							</form>
-						</div>
-					</div>
-				</div>
-
+				if (file_exists($file)) {
+					require $file;
+				}
+				?>
 				<div class="clear"></div>
 
 			  <?php break;
